test(unit): converted tests for the test script builder to vitest
for #452
